+++ nodes/Sleeper/descriptions/index.ts
@@ -1,7 +1,13 @@
 import type { INodeProperties } from 'n8n-workflow';
 
 import { leagueDescription } from './LeagueDescription';
+import { leagueUserDescription } from './LeagueUserDescription';
+import { matchupDescription } from './MatchupDescription';
+import { playoffDescription } from './PlayoffDescription';
+import { rosterDescription } from './RosterDescription';
 import { sportDescription } from './SportDescription';
+import { tradedPickDescription } from './TradedPickDescription';
+import { transactionDescription } from './TransactionDescription';
 import { userDescription } from './UserDescription';
 
 export const sleeperProperties: INodeProperties[] = [
@@ -16,8 +22,32 @@ export const sleeperProperties: INodeProperties[] = [
 				value: 'league',
 			},
 			{
+				name: 'League User',
+				value: 'leagueUser',
+			},
+			{
+				name: 'Matchup',
+				value: 'matchup',
+			},
+			{
+				name: 'Playoff',
+				value: 'playoff',
+			},
+			{
+				name: 'Roster',
+				value: 'roster',
+			},
+			{
 				name: 'Sport',
 				value: 'sport',
+			},
+			{
+				name: 'Traded Pick',
+				value: 'tradedPick',
+			},
+			{
+				name: 'Transaction',
+				value: 'transaction',
 			},
 			{
 				name: 'User',
@@ -28,5 +58,11 @@ export const sleeperProperties: INodeProperties[] = [
 	},
 	...userDescription,
 	...leagueDescription,
+	...leagueUserDescription,
+	...matchupDescription,
+	...playoffDescription,
+	...rosterDescription,
 	...sportDescription,
+	...tradedPickDescription,
+	...transactionDescription,
 ];